[rename/#28] - 임시로 만든 캘린더 page, 컴포넌트로 변경
